% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.integ.fn}
\alias{calc.integ.fn}
\title{Integrand function for calculating \eqn{phi} using integration}
\usage{
calc.integ.fn(r, gx.funct.name = "dgamma", theta = theta, lambda = lambda,
  psi = psi, ...)
}
\arguments{
\item{r}{Spatial distance from an indivual}

\item{gx.funct.name}{a string; function for representing \eqn{g(r)}, the contact distance probability distribution.
The gamma distribution function (i.e. \code{dgamma()}, input as \code{gx.funct.name='dgamma'}. Custom functions, such as splines can also be inputted.)}

\item{theta}{the maximum probability ratio of one susceptible individual being in contact with a susceptible at a distance of 0 (r=0)}

\item{lambda}{rate of decay of the exponential distribution. Describes the dispersion of the \eqn{tau(r)} function}

\item{psi}{a rescaling value, allowing the distribution to average to 1 across the entire range of \code{r}.}

\item{...}{additional arguments to be passed to function described by \code{gx.funct.name}.}
}
\value{
vector of \eqn{tau(r) * g(r)} at all values of \code{r}.
}
\description{
Integrand function for calculating \eqn{phi} using integration
}
\examples{

}
\author{
Shaun Truelove
}
