% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.integ.fn.spline}
\alias{calc.integ.fn.spline}
\title{Integrand function to feed to the R effective and Phi functions}
\usage{
calc.integ.fn.spline(r, tau.fun.spline, gx.funct.name = "dgamma", ...)
}
\arguments{
\item{r}{Spatial distance from an indivual}

\item{tau.fun.spline}{spline function defined for \eqn{tau(r)} function.}

\item{gx.funct.name}{a string; function for representing \eqn{g(r)}, the contact distance probability distribution.
The gamma distribution function (i.e. \code{dgamma()}, input as \code{gx.funct.name='dgamma'}. Custom functions, such as splines can also be inputted.).}

\item{...}{additional arguments to be passed to \code{gx.funct}.}
}
\value{
vector of \eqn{tau(r) * g(r)} at all values of \code{r}.
}
\description{
Integrand function to feed to the R effective and Phi functions
}
\examples{

}
\author{
Shaun Truelove
}
