% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.phi.analytic}
\alias{calc.phi.analytic}
\title{Analytic Calculation of Phi (integral of \eqn{tau(x)*g(x)})}
\usage{
calc.phi.analytic(theta, lambda, upper.lim, lower.lim = 1e-05,
  subdivisions = 1000L, by.x = 0.001, ...)
}
\arguments{
\item{theta}{the maximum probability ratio of one susceptible individual being in contact with a susceptible at a distance of 0 (r=0)}

\item{lambda}{rate of decay of the exponential distribution. Describes the dispersion of the \eqn{tau(r)} function}

\item{upper.lim}{upper limit of integration.}

\item{lower.lim}{lower limit of integration.}

\item{subdivisions}{the maximum number of subintervals in the integration.}

\item{by.x}{r discretization value.}

\item{...}{additional arguments to be passed to \code{calc.integ.fn}.}
}
\value{
\eqn{phi}, the clustering adjustement factor.
}
\description{
Analytic Calculation of Phi (integral of \eqn{tau(x)*g(x)})
}
\examples{

}
\author{
Shaun Truelove
}
