% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.phi.analytic.set.tau}
\alias{calc.phi.analytic.set.tau}
\title{Calculate \eqn{phi} analytically using parametric functions for \eqn{tau(r)} and \eqn{g(r)}}
\usage{
calc.phi.analytic.set.tau(theta, lambda, psi, upper.lim, lower.lim = 1e-06,
  subdivisions = 1000L, ...)
}
\arguments{
\item{theta}{the maximum probability ratio of one susceptible individual being in contact with a susceptible at a distance of 0 (r=0)}

\item{lambda}{rate of decay of the exponential distribution. Describes the dispersion of the \eqn{tau(r)} function}

\item{psi}{a rescaling value, allowing the distribution to average to 1 across the entire range of \code{r}.}

\item{upper.lim}{upper limit of integration.}

\item{lower.lim}{lower limit of integration.}

\item{subdivisions}{the maximum number of subintervals in the integration.}

\item{...}{additional arguments to be passed to \code{calc.integ.fn}.}
}
\value{
\eqn{phi}, the clustering adjustement factor.
}
\description{
Calculate \eqn{phi} analytically using parametric functions for \eqn{tau(r)} and \eqn{g(r)}
}
\examples{

}
\author{
Shaun Truelove
}
