% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.phi.cf}
\alias{calc.phi.cf}
\title{Analytic Calculation of Phi (closed form integration of tau(x)*g(x))}
\usage{
calc.phi.cf(theta, lambda, psi, alpha, beta)
}
\arguments{
\item{theta}{the maximum probability ratio of one susceptible individual being in contact with a susceptible at a distance of 0 (r=0).}

\item{lambda}{rate of decay of the exponential distribution. Describes the dispersion of the \eqn{tau(r)} function.}

\item{psi}{a rescaling value, allowing the distribution to average to 1 across the entire range of \code{r}.}

\item{alpha}{shape parameter from gamma distribution.}

\item{beta}{rate parameter from gamma distribution. Scale parameter = 1/rate.}
}
\value{
\eqn{phi}, the clustering adjustement factor.
}
\description{
tau(r): theta*exp(-lambda*r)+psi
g(r): f(r, alpha, beta) ~ gamma(shape=alpha, rate=beta)
}
\examples{

}
\author{
Shaun Truelove
}
