% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.phi.spline}
\alias{calc.phi.spline}
\title{Calculate \eqn{phi} using spline function for \eqn{tau(r)}}
\usage{
calc.phi.spline(tau.fun.spline, gx.funct.name = "dgamma", upper.lim,
  lower.lim = 1e-06, subdivisions = 1000L, ...)
}
\arguments{
\item{tau.fun.spline}{spline function defined for \eqn{tau(r)} function.}

\item{gx.funct.name}{a string; function for representing \eqn{g(r)}, the contact distance probability distribution.
The gamma distribution function (i.e. \code{dgamma()}, input as \code{gx.funct.name='dgamma'}. Custom functions, such as splines can also be inputted.).}

\item{upper.lim}{upper limit of integration.}

\item{lower.lim}{lower limit of integration.}

\item{subdivisions}{the maximum number of subintervals in the integration.}

\item{...}{additional arguments to be passed to \code{calc.intTeg.fn.spline}.}
}
\value{
\eqn{phi}, the clustering adjustement factor.
}
\description{
Calculate \eqn{phi} using spline function for \eqn{tau(r)}
}
\examples{

}
\author{
Shaun Truelove
}
