% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.R.analytic.phi}
\alias{calc.R.analytic.phi}
\title{Calculate R analytically from known phi}
\usage{
calc.R.analytic.phi(phi, v, VE = 1, cbeta = c(10, 15, 20))
}
\arguments{
\item{phi}{Clustering adjustment factor, \eqn{phi}. This is the relative probability of contact between susceptibles resulting from heterogeneity, compared to a homogenious population.}

\item{v}{vaccination coverage.}

\item{VE}{vaccine efficacy. Default is VE=1, assuming we are looking at successful vaccination only (i.e. critical immunization threshold).}

\item{cbeta}{equivalent to \eqn{R[0]}, represents contact number per infectious period (c) x transmission probability per contact (beta).}
}
\value{
\eqn{R} effective, adjusted for spatial clustering of susceptibility
}
\description{
Calculate R analytically from known phi
}
\author{
Shaun Truelove
}
