% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.R.cf}
\alias{calc.R.cf}
\title{Calculate R analytically with the closed-form equation, assuming an exponential form of \code{tau(r)} and a gamma distribution for \code{g(r)}}
\usage{
calc.R.cf(v, VE = 1, cbeta = c(10, 15, 20), theta, lambda, psi, alpha, beta)
}
\arguments{
\item{v}{vaccination coverage.}

\item{VE}{vaccine efficacy. Default is VE=1, assuming we are looking at successful vaccination only (i.e. critical immunization threshold).}

\item{cbeta}{equivalent to \eqn{R[0]}, represents contact number per infectious period (c) x transmission probability per contact (beta).}

\item{theta}{the maximum probability ratio of one susceptible individual being in contact with a susceptible at a distance of 0 (r=0)}

\item{lambda}{rate of decay of the exponential distribution. Describes the dispersion of the \eqn{tau(r)} function}

\item{psi}{a rescaling value, allowing the distribution to average to 1 across the entire range of \code{r}.}

\item{alpha}{shape parameter from gamma distribution.}

\item{beta}{rate parameter from gamma distribution. Scale parameter = 1/rate.}

\item{upper.lim}{upper limit of integration.}
}
\value{
\eqn{R} effective, adjusted for spatial clustering of susceptibility
}
\description{
Calculate R analytically with the closed-form equation, assuming an exponential form of \code{tau(r)} and a gamma distribution for \code{g(r)}
}
\examples{

}
\author{
Shaun Truelove
}
