% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.R}
\alias{calc.R}
\title{Calculate Traditional, Homogeneous R Effective from VE, v, and R0}
\usage{
calc.R(Bc, VE, v)
}
\arguments{
\item{Bc}{Bc = R0 for the particular agent.}

\item{VE}{vaccine efficacy. Default is VE=1, assuming we are looking at successful vaccination only (i.e. critical immunization threshold).}

\item{v}{vaccination coverage. vaccination coverage.}
}
\value{
traditional, homogeneous \eqn{R} effective estimate.
}
\description{
Calculate Traditional, Homogeneous R Effective from VE, v, and R0
}
\examples{

}
\author{
Shaun Truelove
}
