% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.R0.from.Vc.analytic}
\alias{calc.R0.from.Vc.analytic}
\title{Calculate \eqn{R[0]} from \eqn{V[c]}}
\usage{
calc.R0.from.Vc.analytic(VE = 1, v = 0.95, theta, lambda, upper.lim, ...)
}
\arguments{
\item{VE}{vaccine efficacy. Default is VE=1, assuming we are looking at successful vaccination only (i.e. critical immunization threshold).}

\item{v}{vaccination coverage.}

\item{theta}{the maximum probability ratio of one susceptible individual being in contact with a susceptible at a distance of 0 (r=0).}

\item{lambda}{rate of decay of the exponential distribution. Describes the dispersion of the \eqn{tau(r)} function.}

\item{upper.lim}{upper limit of integration.}

\item{...}{additional arguments to be passed to \code{calc.integ.fn}.}
}
\value{
\eqn{R[0]}
}
\description{
Analytic version using integration to back-calculate \eqn{R[0]} from \eqn{V[c]} for diagnostic and graphical purposes.
Here the \eqn{g(r)} function must be specified along with its parameters (i.e. a gamma distribution is not assumed).
An exponential distribution for \eqn{tau(r)} is assumed.
}
\examples{

}
\author{
Shaun Truelove
}
