% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.R0.from.Vc}
\alias{calc.R0.from.Vc}
\title{Calculate \eqn{R[0]} from \eqn{V[c]}}
\usage{
calc.R0.from.Vc(VE = 1, Vc = 0.95, phi)
}
\arguments{
\item{VE}{vaccine efficacy. Default is VE=1, assuming we are looking at successful vaccination only.}

\item{Vc}{Critical vaccination threshold (\eqn{V[c]}), a.k.a critical immunity threshold (\eqn{I[c]}) when \eqn{VE=1}.}

\item{phi}{Clustering adjustment factor, \eqn{phi}. This is the relative probability of contact between susceptibles resulting from heterogeneity, compared to a homogenious population. Clustering adjustment factor, \eqn{phi}. This is the relative probability of contact between susceptibles resulting from heterogeneity, compared to a homogenious population.}
}
\value{
\eqn{R[0]}
}
\description{
Back-calculate the \eqn{R[0]} from known \eqn{V[c]} and \eqn{phi}
}
\examples{

}
\author{
Shaun Truelove
}
