% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.tau.analytic}
\alias{calc.tau.analytic}
\title{Function to produce analytic \eqn{tau(r)} data for plots}
\usage{
calc.tau.analytic(theta, lambda = 0.5, max.x = 50, r.delim = 1)
}
\arguments{
\item{theta}{the maximum probability ratio of one susceptible individual being in contact with a susceptible at a distance of 0 (r=0)}

\item{lambda}{rate of decay of the exponential distribution. Describes the dispersion of the \eqn{tau(r)} function}

\item{r.delim}{\code{r} distance window size (in km) for discrete distance ranges.}

\item{upper.lim}{upper limit of integration.}
}
\value{
vector of \eqn{tau(r)} for discrete \code{r} distance window ranges.
}
\description{
Function to produce analytic \eqn{tau(r)} data for plots
}
\examples{

}
\author{
Shaun Truelove
}
