% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.tau.distrib}
\alias{calc.tau.distrib}
\title{Calculate tau(r) for r midpoints}
\usage{
calc.tau.distrib(tau.level = 1.5, lambda = 0.5, max.x = 50,
  r.delim = 0.5)
}
\arguments{
\item{tau.level}{Maximum relative probability of another susceptible individual, occurring at distance r=0.
For example if the overall probability of an individual being susceptible is 0.10 (v=90%),
if tau.level=1.5, the average maximum probability is 0.15, due to clustering.}

\item{lambda}{rate of decay of the exponential distribution. Describes the dispersion of the \eqn{tau(r)} function}

\item{max.x}{maximum distance value for which to calculate the tau vector.}

\item{r.delim}{\code{r} distance window size (in km) for discrete distance ranges.}
}
\value{
vector of \eqn{tau(r)} for discrete \code{r} distance window ranges.
}
\description{
Calculate tau(r) for r midpoints
}
\examples{

}
\author{
Shaun Truelove
}
