% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.tau}
\alias{calc.tau}
\title{Calculate exponential \eqn{tau(r)} for specific value of \eqn{r}, assuming knowledge of maximum contact probability ratio}
\usage{
calc.tau(r, tau.level = 1.5, lambda = 0.5, max.x = 50)
}
\arguments{
\item{r}{Spatial distance from an indivual.}

\item{lambda}{rate of decay of the exponential distribution. Describes the dispersion of the \eqn{tau(r)} function.}

\item{max.x}{maximum distance value for which to calculate the tau vector.}
}
\value{
\eqn{tau(r)} for specific \code{r} distances.
}
\description{
Calculate exponential \eqn{tau(r)} for specific value of \eqn{r}, assuming knowledge of maximum contact probability ratio
}
\examples{

}
\author{
Shaun Truelove
}
