% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.Vc.analytic.params}
\alias{calc.Vc.analytic.params}
\title{Calculate Vc analytically, using integration and previously refined \code{tau(r)} parameters}
\usage{
calc.Vc.analytic.params(VE = 1, R0 = 15, theta, lambda, psi, upper.lim, ...)
}
\arguments{
\item{VE}{vaccine efficacy. Default is VE=1, assuming we are looking at successful vaccination only (i.e. critical immunization threshold).}

\item{R0}{basic reproductive number. Default is \eqn{R[0]=15}, for measles.}

\item{theta}{the maximum probability ratio of one susceptible individual being in contact with a susceptible at a distance of 0 (r=0).}

\item{lambda}{rate of decay of the exponential distribution. Describes the dispersion of the \eqn{tau(r)} function}

\item{psi}{a rescaling value, allowing the distribution to average to 1 across the entire range of \code{r}.}

\item{upper.lim}{upper limit of integration.}

\item{...}{additional arguments to be passed to \code{calc.integ.fn}.}
}
\value{
\eqn{V[c]}, adjusted for spatial clustering of susceptibility
}
\description{
Calculate Vc analytically, using integration and previously refined \code{tau(r)} parameters
}
\examples{

}
\author{
Shaun Truelove
}
