% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.Vc.analytic.phi}
\alias{calc.Vc.analytic.phi}
\title{Calculate Vc analytically, from phi}
\usage{
calc.Vc.analytic.phi(VE = 1, R0 = 15, phi)
}
\arguments{
\item{VE}{vaccine efficacy. Default is VE=1, assuming we are looking at successful vaccination only (i.e. critical immunization threshold).}

\item{R0}{basic reproductive number. Default is \eqn{R[0]=15}, for measles.}

\item{phi}{Clustering adjustment factor, \eqn{phi}. This is the relative probability of contact between susceptibles resulting from heterogeneity, compared to a homogenious population. Clustering adjustment factor, \eqn{phi}. This is the relative probability of contact between susceptibles resulting from heterogeneity, compared to a homogenious population.}
}
\value{
\eqn{V[c]}, adjusted for spatial clustering of susceptibility
}
\description{
Calculate Vc analytically, from phi
}
\author{
Shaun Truelove
}
