% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.Vc.analytic}
\alias{calc.Vc.analytic}
\title{Calculate Vc analytically, using integration}
\usage{
calc.Vc.analytic(VE = 1, R0 = 15, theta, lambda, upper.lim,
  lower.lim = 1e-05, subdivisions = 1000L, r.delim = 0.001, ...)
}
\arguments{
\item{VE}{vaccine efficacy. Default is VE=1, assuming we are looking at successful vaccination only (i.e. critical immunization threshold).}

\item{R0}{basic reproductive number. Default is \eqn{R[0]=15}, for measles.}

\item{theta}{the maximum probability ratio of one susceptible individual being in contact with a susceptible at a distance of 0 (r=0)}

\item{lambda}{rate of decay of the exponential distribution. Describes the dispersion of the \eqn{tau(r)} function}

\item{upper.lim}{upper limit of integration.}

\item{lower.lim}{lower limit of integration.}

\item{subdivisions}{the maximum number of subintervals in the integration.}

\item{r.delim}{r discretization value.}

\item{...}{additional arguments to be passed to \code{calc.integ.fn}.}
}
\value{
\eqn{V[c]}, adjusted for spatial clustering of susceptibility
}
\description{
Calculate Vc analytically, using integration
}
\examples{

}
\author{
Shaun Truelove
}
