% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{calc.Vc.cf}
\alias{calc.Vc.cf}
\title{Calculate Vc analytically using the closed form solution}
\usage{
calc.Vc.cf(VE = 1, R0 = 15, theta, lambda, psi, alpha, beta)
}
\arguments{
\item{VE}{vaccine efficacy. Default is VE=1, assuming we are looking at successful vaccination only (i.e. critical immunization threshold).}

\item{R0}{basic reproductive number. Default is \eqn{R[0]=15}, for measles.}

\item{theta}{the maximum probability ratio of one susceptible individual being in contact with a susceptible at a distance of 0 (r=0)}

\item{lambda}{rate of decay of the exponential distribution. Describes the dispersion of the \eqn{tau(r)} function}

\item{psi}{a rescaling value, allowing the distribution to average to 1 across the entire range of \code{r}.}

\item{alpha}{shape parameter from gamma distribution.}

\item{beta}{rate parameter from gamma distribution. Scale parameter = 1/rate.}
}
\value{
\eqn{V[c]}, adjusted for spatial clustering of susceptibility
}
\description{
THis function uses the closed-form solution of the integral of the exponential \code{tau(r)} function and gamma \code{g(r)} function.
}
\examples{

}
\author{
Shaun Truelove
}
