% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_dist_params_data.R
\docType{data}
\name{g_params}
\alias{g_params}
\title{Contact distance probability distributions}
\format{A list of 3 lists}
\source{
http://rspb.royalsocietypublishing.org/content/281/1785/20140268.short

 http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0037027

 http://rsos.royalsocietypublishing.org/content/4/5/170046
}
\usage{
data(contact_dist_params)
}
\description{
Gamma distribution parameters for probability distributions parameterized from empirical data from
China (Read et al. 2014), Zambia (Searle et al. 2017), and the U.S. (Noulas et al. 2012).
These data feed the analytic R effective and Ic equations.
}
\references{
Read JM, Lessler J, Riley S, Wang S, Tan LJ, Kwok KO, Guan Y, Jiang CQ, Cummings DAT. 2014
Social mixing patterns in rural and urban areas of southern China. Proc. Biol. Sci. 281, 20140268. (doi:10.1098/rspb.2014.0268)

Noulas A, Scellato S, Lambiotte R, Pontil M, Mascolo C. 2012
A Tale of Many Cities: Universal Patterns in Human Urban Mobility. PLoS ONE 7, e37027. (doi:10.1371/journal.pone.0037027)

Searle KM, Lubinda J, Hamapumbu H, Shields TM, Curriero FC, Smith DL, Thuma PE, Moss WJ. 2017
Characterizing and quantifying human movement patterns using GPS data loggers in an area approaching malaria elimination in rural southern Zambia. R. Soc. Open Sci. 4, 170046. (doi:10.1098/rsos.170046)
}
\keyword{data}
