% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates_source.R
\name{get.tau.params}
\alias{get.tau.params}
\title{Get \eqn{tau(r)} exponential function parameters for use in other functions}
\usage{
get.tau.params(tau.level = 1.5, lambda = 0.5, max.x = 50, r.delim = 1)
}
\arguments{
\item{tau.level}{Maximum relative probability of another susceptible individual, occurring at distance r=0.
For example if the overall probability of an individual being susceptible is 0.10 (v=90%),
if tau.level=1.5, the average maximum probability is 0.15, due to clustering.}

\item{lambda}{rate of decay of the exponential distribution. Describes the dispersion of the \eqn{tau(r)} function.}

\item{max.x}{maximum distance value for which to calculate the tau vector.}

\item{r.delim}{\code{r} distance window size (in km) for discrete distance ranges.}
}
\value{
list of \eqn{tau(r)} function parameters:
  \item{theta}{maximum probability ratio of one susceptible individual being in contact with a susceptible at a distance of 0 (r=0).}
  \item{lambda}{rate of decay of the exponential distribution. Describes the dispersion of the \eqn{tau(r)} function.}
  \item{psi}{a rescaling value, allowing the distribution to average to 1 across the entire range of \code{r}.}
}
\description{
Get \eqn{tau(r)} exponential function parameters for use in other functions
}
\examples{

}
\author{
Shaun Truelove
}
